refactor: 1. 파일 구조 정리 2. 클래스명 컨벤션 맞춰 수정 3. 사용하지 않는 import삭제
